Backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-gis==1.0
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
python-decouple==3.8
django-filter==24.2